requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose==3.3.0
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
